Add a simple test to serve as a bit of sanity check
This is the first unit test in this project, so I wanted to raise this
very simple one to see how the maintainers feel about it.

This runs fine and correctly. It does not use any database connection.

The aim here was to use the setup-test-assert principle, with some of
the noiser and repeated shared set up pulled out to the `setUp` method.